split proxyWallet.ts into relayExecution + tradingApprovals
